tidak dapat menghapus kua

+++ app/Http/Livewire/Kua/Form.php
@@ -78,9 +78,16 @@ class Form extends Kua
 
     public function destroy()
     {
-        $kua = ModelsKua::findOrFail($this->kuaIdDelete);
+        $kua = ModelsKua::withCount('penghulus')->findOrFail($this->kuaIdDelete);
 
         $this->authorize('delete', $kua);
+
+        // Jika KUA memiliki penghulu
+        if ($kua->penghulus_count > 0) {
+            session()->flash('message', 'Data KUA ' . $kua->name .' tidak dapat dihapus');
+            $this->closeModal();
+            return redirect('kua');
+        }
 
         $kua->delete();
         session()->flash('message', 'Data KUA ' . $kua->name .' berhasil dihapus');
